Nove schema entities out of top-level
